Refactor test directory (#403)
* Flatten test directory and unify one-file-per-module convention

Closes #381.

- Move test/syntax/*.py and test/semantics/*.py up to test/, so each
  discopy/<module>.py again has a matching test/<module>.py.
- Move test/utils/utils.py to test/utils.py and relocate the shared test
  data (pickles, zx-graph.json) to a new test/fixtures/ folder.
- Merge test/quantum/eval.py into test/quantum/circuit.py so the quantum
  submodule keeps one test file per module (Circuit.eval lives in circuit).
- Update pyproject.toml testpaths/ignore and the paths referenced from
  test/utils.py and test/quantum/zx.py, plus the AGENTS.md description.


* Clean up test/fixtures/pickles per PR review

Addresses review feedback on #403:
- Remove the pickle TODO/ folder (stubs, not used by any test)
- Remove the src/ folder and the now-orphaned test_pickle_unpickle
- Remove generate_pickles.py (fixtures are trivially regenerable)
- Remove the unreferenced rigid.Box.legacy.pickle duplicate
- Rename main/ -> 1.X/ to reflect the version it holds
- Rename quantum.gates.pkl -> .pickle in both 1.X/ and 0.6/

Point test_pickle_version_compatibility at 1.X/ and drop the now-unused
sys/pkgutil imports.


* Address review round 2: version dirs, python submodule split, merge main

- Rename pickle fixture dir 1.X/ -> 1.3/ (the unreleased breaking-changes
  version) and add a 1.2/ dir with the same fixtures regenerated under
  discopy 1.2.2 (the latest release).
- test_pickle_version_compatibility now checks the 1.3 reference against
  both older versions (0.6 and 1.2).
- Split test/python.py into one file per submodule under test/python/
  (multiplicative, finset, additive), matching the grammar/quantum/drawing
  convention; note python in AGENTS.md's test-layout description.

Also merged origin/main to pick up the Functor ob/ar -> ob_map/ar_map
rename (#369).


* Make pickle version-compat test tolerant to float round-off

The 1.2 fixtures are generated under discopy 1.2.2, whose numpy computes the
Hadamard gate's 1 / sqrt(2) entries with a last-bit difference from the numpy
that produced the 0.6/1.3 fixtures. Exact pickle equality is therefore not
portable across versions for gate matrices; compare reprs with floats rounded
to 12 significant figures, which still catches any real structural difference.


* Regenerate docs/_static

---------
