+++ test/compact.py
