+++ test/fixtures/pickles/0.6/pregroup.Word.pickle
