+++ test/fixtures/pickles/0.6/quantum.Qudit.pickle
