+++ test/fixtures/pickles/0.6/quantum.gates.pickle
