+++ test/fixtures/pickles/0.6/quantum.zx.Diagram.pickle
