+++ test/fixtures/pickles/0.6/rigid.Box.pickle
