+++ test/fixtures/pickles/0.6/tensor.Box.pickle
