+++ test/fixtures/pickles/1.3/quantum.Qudit.pickle
