+++ test/monoidal.py
