+++ test/stream.py
