--- test/utils/pickles/src/quantum_Qudit.py
@@ -1,5 +1,0 @@
-from discopy.quantum.circuit import Qudit, Ty
-
-
-pick = [Qudit(d) for d in range(2, 5)]
-pick += list(map(Ty, pick))
