--- test/utils/pickles/src/rigid_Box.py
@@ -1,3 +1,0 @@
-from discopy.rigid import *
-
-pick = Box('f', Ty('x', 'x'), Ty('y'))
